set room pic feature. also update and decommission a room

+++ RFS/WebApp/Controllers/ApiControllers/RoomsController.cs
@@ -1,10 +1,14 @@
-﻿using RFS.Models.Entities;
+﻿using RFS.Models;
+using RFS.Models.Entities;
 using RoomManagement;
 using RoomManagement.Entities;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Net;
 using System.Net.Http;
+using System.Threading.Tasks;
 using System.Web;
 using System.Web.Http;
 using System.Web.Mvc;
@@ -16,11 +20,14 @@ namespace RFS_API.Controllers
         IRoomManager _roomManager = null;
         IBookingManager _bookingManager = null;
         ILocationManager _locationManager = null;
-        public RoomsController(IRoomManager roomManager, IBookingManager bookingManager,ILocationManager locationManager)
+        IUserManager _userManager = null;
+
+        public RoomsController(IUserManager userManager,IRoomManager roomManager, IBookingManager bookingManager,ILocationManager locationManager)
         {
             _roomManager = roomManager;
             _bookingManager = bookingManager;
             _locationManager = locationManager;
+            _userManager = userManager;
         }
         // GET api/<controller>
         public IEnumerable<Room> Get()
@@ -34,9 +41,80 @@ namespace RFS_API.Controllers
         public IEnumerable<Room> GetRoomsUnderLocation(string locationId)
         {
 
-            return _roomManager.GetAllRoomsForLocation(locationId);
+            return _roomManager.GetAllRoomsForLocation(locationId).Select(x=> new Room() {Id=x.Id,decommission = x.decommission, location= x.location, MonitorScreen = x.MonitorScreen,  Projector = x.Projector, RoomName = x.RoomName,  Sitting = x.Sitting, VideoConferencing_ = x.VideoConferencing_ });
         }
 
+        [System.Web.Http.HttpPost]
+        [System.Web.Http.Route("api/rooms/{roomid}/dp")]
+        public async Task<IHttpActionResult> SetDP(string roomid)
+        {
+            var useremail = new TApiAuth().GetLoggedInUsername(Request);
+            if (string.IsNullOrEmpty(useremail))
+            {
+                return BadRequest();
+            }
+            var user = _userManager.GetUserFromMailId(useremail);
+            if (user.isAdmin.HasValue && user.isAdmin.Value)
+            {
+                if (!Request.Content.IsMimeMultipartContent())
+                    throw new HttpResponseException(HttpStatusCode.UnsupportedMediaType);
+
+                var provider = new MultipartMemoryStreamProvider();
+                await Request.Content.ReadAsMultipartAsync(provider);
+                foreach (var file in provider.Contents)
+                {
+                    var filename = file.Headers.ContentDisposition.FileName.Trim('\"');
+                    var buffer = await file.ReadAsByteArrayAsync();
+                    _roomManager.SetRoomProfilePic(roomid, buffer, filename);
+                }
+
+                return Ok();
+            }
+            else
+            {
+                return Unauthorized();
+            }
+        }
+
+        [System.Web.Http.HttpGet]
+        [System.Web.Http.Route("api/rooms/{roomid}/dp")]
+        public HttpResponseMessage GetDP(string roomid)
+        {
+            var useremail = new TApiAuth().GetLoggedInUsername(Request);
+            if (string.IsNullOrEmpty(useremail))
+            {
+                return new HttpResponseMessage(HttpStatusCode.Unauthorized);
+            }
+
+            var room = _roomManager.GetRoomById(roomid);
+            if (room == null)
+            {
+                return new HttpResponseMessage(HttpStatusCode.BadRequest);
+            }
+            HttpResponseMessage response = new HttpResponseMessage(HttpStatusCode.OK);
+            var profilepic = room.RoomProfilePics.FirstOrDefault();
+            Byte[] b;
+            if (profilepic == null)
+            {
+                string path = HttpContext.Current.Server.MapPath("~\\Content\\img\\room-outline.png");
+                int a = 1;
+                b = File.ReadAllBytes(path);
+                //using (FileStream fs = new FileStream(path, FileMode.Open))
+                //{
+                //    response.Content = new StreamContent(fs);
+                //}
+
+            }
+            else
+            {
+                b = profilepic.data;
+
+
+            }
+            response.Content = new StreamContent(new MemoryStream(b));
+            response.Content.Headers.ContentType = new System.Net.Http.Headers.MediaTypeHeaderValue("image/png");
+            return response;
+        }
         [System.Web.Http.Route("api/location/{locationId}/searchrooms")]
         [System.Web.Http.HttpGet()]
         public HttpResponseMessage GetsAvailableRoomsUnderLocation(HttpRequestMessage httpRequest, string locationId,string SdateTime, string EdateTime)
@@ -57,7 +135,8 @@ namespace RFS_API.Controllers
             var rooms = _roomManager.GetAllRoomsForLocation(locationId);
             foreach (var r in rooms)
             {
-
+                if (r.decommission.HasValue && r.decommission.Value)
+                    continue;
                 if (_bookingManager.GetBookingForRoom(r.Id, s, e).Count() == 0)
                 {
                     var room = new { Id = r.Id, location = r.location, MonitorScreen = r.MonitorScreen, Projector = r.Projector, RoomName= r.RoomName, Sitting=r.Sitting, VideoConferencing=r.VideoConferencing_,image = "Content\\img\\room.jpg" };
@@ -88,15 +167,41 @@ namespace RFS_API.Controllers
         }
 
         // PUT api/<controller>/5
-        public void Put(string id, [FromBody]Room room)
+        public IHttpActionResult Put(string id, [FromBody]Room room)
         {
-            _roomManager.UpdateRoomProperties(id, room);
+            var useremail = new TApiAuth().GetLoggedInUsername(Request);
+
+            if (!string.IsNullOrEmpty(useremail))
+            {
+
+                user u = _userManager.GetUserFromMailId(useremail);
+                if (u != null && u.isAdmin.HasValue && u.isAdmin.Value)
+                    _roomManager.UpdateRoomProperties(id, room);
+                return Ok();
+            }
+            else
+            {
+                return Unauthorized();
+            }
         }
 
         // DELETE api/<controller>/5
-        public void Delete(string id)
+        public IHttpActionResult Delete(string id)
         {
-            _roomManager.DeleteRoom(id);
+            var useremail = new TApiAuth().GetLoggedInUsername(Request);
+            
+            if (!string.IsNullOrEmpty(useremail)){
+
+                user u = _userManager.GetUserFromMailId(useremail);
+                if(u!=null && u.isAdmin.HasValue && u.isAdmin.Value)
+                _roomManager.DeleteRoom(id);
+                return Ok();
+            }
+            else
+            {
+                return Unauthorized();
+            }
+
         }
     }
 }
